+++ src/analysis/modals/ComputeModal/GcpComputeModal/GcpComputeModal.js
@@ -47,9 +47,9 @@ import { ClipboardButton } from 'src/components/ClipboardButton';
 import { ButtonOutline, ButtonPrimary, IdContainer, LabeledCheckbox, Link, Select, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
 import { ImageDepViewer } from 'src/components/ImageDepViewer';
+import { InfoBox } from 'src/components/InfoBox';
 import { NumberInput, TextInput, ValidatedInput } from 'src/components/input';
 import { withModalDrawer } from 'src/components/ModalDrawer';
-import { InfoBox } from 'src/components/PopupTrigger';
 import { getAvailableComputeRegions, getLocationType, getRegionInfo, isLocationMultiRegion, isUSLocation } from 'src/components/region-common';
 import TitleBar from 'src/components/TitleBar';
 import { Ajax } from 'src/libs/ajax';
