+++ src/components/group-common.js
@@ -3,10 +3,11 @@ import { Fragment, useState } from 'react';
 import { b, div, h, label } from 'react-hyperscript-helpers';
 import { ButtonPrimary, IdContainer, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
 import { icon } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import { AutocompleteTextInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import Modal from 'src/components/Modal';
-import { InfoBox, makeMenuIcon, MenuTrigger } from 'src/components/PopupTrigger';
+import { makeMenuIcon, MenuTrigger } from 'src/components/PopupTrigger';
 import { ariaSort, HeaderRenderer } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
 import { Ajax } from 'src/libs/ajax';
