+++ src/pages/Profile.js
@@ -4,9 +4,9 @@ import { div, h, h2, h3, label, p, span } from 'react-hyperscript-helpers';
 import { ClipboardButton } from 'src/components/ClipboardButton';
 import { ButtonPrimary, Checkbox, IdContainer, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
 import FooterWrapper from 'src/components/FooterWrapper';
+import { InfoBox } from 'src/components/InfoBox';
 import { TextInput, ValidatedInput } from 'src/components/input';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
-import { InfoBox } from 'src/components/PopupTrigger';
 import ProfilePicture from 'src/components/ProfilePicture';
 import { SimpleTabBar } from 'src/components/tabBars';
 import TopBar from 'src/components/TopBar';
