+++ src/pages/billing/Project.js
@@ -6,10 +6,11 @@ import { cloudProviders } from 'src/analysis/utils/runtime-utils';
 import { ButtonPrimary, customSpinnerOverlay, IdContainer, Link, Select } from 'src/components/common';
 import { DeleteUserModal, EditUserModal, MemberCard, MemberCardHeaders, NewUserCard, NewUserModal } from 'src/components/group-common';
 import { icon } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import { TextInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import Modal from 'src/components/Modal';
-import { InfoBox, MenuTrigger } from 'src/components/PopupTrigger';
+import { MenuTrigger } from 'src/components/PopupTrigger';
 import { SimpleTabBar } from 'src/components/tabBars';
 import { ariaSort, HeaderRenderer } from 'src/components/table';
 import { Ajax } from 'src/libs/ajax';
