+++ src/pages/profile/external-identities/NihLink.ts
@@ -5,8 +5,8 @@ import { div, h, h3, span } from 'react-hyperscript-helpers';
 import Collapse from 'src/components/Collapse';
 import { ButtonPrimary, Link, spinnerOverlay } from 'src/components/common';
 import { ShibbolethLink } from 'src/components/external-account-links';
+import { InfoBox } from 'src/components/InfoBox';
 import Modal from 'src/components/Modal';
-import { InfoBox } from 'src/components/PopupTrigger';
 import { Ajax } from 'src/libs/ajax';
 import { NihDatasetPermission } from 'src/libs/ajax/User';
 import { withErrorReporting } from 'src/libs/error';
