+++ src/pages/workspaces/workspace/Dashboard.js
@@ -7,8 +7,8 @@ import { ClipboardButton } from 'src/components/ClipboardButton';
 import Collapse from 'src/components/Collapse';
 import { ButtonPrimary, ButtonSecondary, LabeledCheckbox, Link, spinnerOverlay } from 'src/components/common';
 import { centeredSpinner, icon, spinner } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import { MarkdownEditor, MarkdownViewer } from 'src/components/markdown';
-import { InfoBox } from 'src/components/PopupTrigger';
 import { getRegionInfo } from 'src/components/region-common';
 import RequesterPaysModal from 'src/components/RequesterPaysModal';
 import { SimpleTable, TooltipCell } from 'src/components/table';
