+++ src/pages/workspaces/workspace/workflows/LaunchAnalysisModal.js
@@ -3,9 +3,9 @@ import { Fragment, useState } from 'react';
 import { b, div, h, label, p, span } from 'react-hyperscript-helpers';
 import { ButtonPrimary, IdContainer, Link } from 'src/components/common';
 import { icon, spinner } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import { ValidatedTextArea } from 'src/components/input';
 import Modal from 'src/components/Modal';
-import { InfoBox } from 'src/components/PopupTrigger';
 import { getRegionInfo } from 'src/components/region-common';
 import { Ajax } from 'src/libs/ajax';
 import { launch } from 'src/libs/analysis';
