+++ src/pages/workspaces/workspace/workflows/WorkflowView.js
@@ -17,11 +17,12 @@ import {
 } from 'src/components/common';
 import Dropzone from 'src/components/Dropzone';
 import { centeredSpinner, icon } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import { DelayedAutocompleteTextArea, DelayedSearchInput, NumberInput } from 'src/components/input';
 import { MarkdownViewer } from 'src/components/markdown';
 import { MenuButton } from 'src/components/MenuButton';
 import Modal from 'src/components/Modal';
-import { InfoBox, makeMenuIcon, MenuTrigger } from 'src/components/PopupTrigger';
+import { makeMenuIcon, MenuTrigger } from 'src/components/PopupTrigger';
 import StepButtons from 'src/components/StepButtons';
 import { HeaderCell, SimpleFlexTable, SimpleTable, Sortable, TextCell } from 'src/components/table';
 import TooltipTrigger from 'src/components/TooltipTrigger';
