+++ src/workflows-app/components/LogViewer.ts
@@ -4,8 +4,8 @@ import { useCallback, useEffect, useState } from 'react';
 import { dd, div, dl, dt, h, span } from 'react-hyperscript-helpers';
 import { ButtonOutline } from 'src/components/common';
 import { centeredSpinner, icon } from 'src/components/icons';
+import { InfoBox } from 'src/components/InfoBox';
 import Modal from 'src/components/Modal';
-import { InfoBox } from 'src/components/PopupTrigger';
 import { SimpleTabBar } from 'src/components/tabBars';
 import { Ajax } from 'src/libs/ajax';
 import { useCancellation } from 'src/libs/react-utils';
